Implementation for managing and getting the words

+++ src/main/resources/migrations/V1__Initial_schema_setup.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS word (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     text VARCHAR(255) NOT NULL,
     type VARCHAR(20) CHECK (type IN ('SUBJECT', 'VERB', 'OBJECT')) NOT NULL,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
@@ -10,10 +10,10 @@ CREATE INDEX idx_word_type ON word(type);
 CREATE INDEX idx_word_text ON word(text);
 
 CREATE TABLE IF NOT EXISTS allowed_combination (
-    id SERIAL PRIMARY KEY,
-    subject_id INT NOT NULL REFERENCES word(id) ON DELETE CASCADE,
-    verb_id INT NOT NULL REFERENCES word(id) ON DELETE CASCADE,
-    object_id INT NOT NULL REFERENCES word(id) ON DELETE CASCADE,
+    id BIGSERIAL PRIMARY KEY,
+    subject_id BIGINT NOT NULL REFERENCES word(id) ON DELETE CASCADE,
+    verb_id BIGINT NOT NULL REFERENCES word(id) ON DELETE CASCADE,
+    object_id BIGINT NOT NULL REFERENCES word(id) ON DELETE CASCADE,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     UNIQUE (subject_id, verb_id, object_id),
     CONSTRAINT fk_subject FOREIGN KEY (subject_id) REFERENCES word(id),
